+++ test-data/variantnormalizer-test/01_IN_altered_multiallelic.vcf
